Ensured time is processed as doubles. DIKERNEL-366

+++ Src/DiKErnel.Integration/Data/AsphaltRevetmentWaveImpact/AsphaltRevetmentWaveImpactLocationDependentInput.cs
@@ -195,7 +195,7 @@ namespace DiKErnel.Integration.Data.AsphaltRevetmentWaveImpact
             double incrementDamage = AsphaltRevetmentWaveImpactFunctions.IncrementDamage(input);
             double damage = RevetmentFunctions.Damage(incrementDamage, initialDamage);
 
-            int? timeOfFailure = null;
+            double? timeOfFailure = null;
 
             if (RevetmentFunctions.FailureRevetment(damage, initialDamage, FailureNumber))
             {
